modify enterprise details list

+++ ehrest/src/main/java/com/everhomes/techpark/expansion/ListEnterpriseDetailResponse.java
@@ -5,16 +5,16 @@ import java.util.List;
 import com.everhomes.discover.ItemType;
 
 public class ListEnterpriseDetailResponse {
-    private Integer nextPageAnchor;
+    private Long nextPageAnchor;
     
     @ItemType(EnterpriseDetailDTO.class)
     private List<EnterpriseDetailDTO> details;
 
-    public Integer getNextPageAnchor() {
+    public Long getNextPageAnchor() {
         return nextPageAnchor;
     }
 
-    public void setNextPageAnchor(Integer nextPageAnchor) {
+    public void setNextPageAnchor(Long nextPageAnchor) {
         this.nextPageAnchor = nextPageAnchor;
     }
